requirements.txt
Flask==3.0.3
blinker==1.8.2
Werkzeug==3.0.3
click==8.1.7
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==2.1.5
paho-mqtt==2.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
typing-extensions==4.11.0
greenlet==3.0.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-socketio==5.11.2
python-engineio==4.9.0
bidict==0.23.1
simple-websocket==1.0.0
wsproto==1.2.0
h11==0.14.0
Flask-WTF==1.2.1
WTForms==3.1.2
requests==2.31.0
urllib3==2.2.1
idna==3.7
charset-normalizer==3.3.2
certifi==2024.2.2
Flask-Moment==1.0.5
packaging==24.0
pymysql==1.1.0
python-dotenv==1.0.1
Flask-Cors==4.0.1